Flake Finder plugin for pytest.
Summary:
This pytest plugin will run all your selected tests multiple
times.  It defaults to 50 but it can be modified to do more or less.

You can combine it with -k to select only some tests, --trace-logging
to make it spew traces, and -x to make it quit when it find one bad
test.

Test Plan:
Created a flaky test, ran it and it found it.  Tested all
the command line arguments.