Fix flakiness in flake finder plugin.
Summary:
Yo Dawg!  I heard you hate flaky tests ...

If you run --flake-finder on a parametrized tests you had a high
likelihood of failing some of them.  This was rare and would never
happen if you ran -n0.  This is actually because the failure happens
only if you run the same parametrized test twice in a row.  And the
failure happens because when you run the the 2nd time pytest won't
call the .setup() on the Function object. (It seems to think that it's
already done it)

Anyway pytest details aside, the solution is to use
pytest_generate_tests to actually generate the flakes.  This only
works for real pytest tests not for unittest.TestCase so we still use
the item multiplication for those.  unittest.TestCase go through a
different setUp process and don't fail if you run them multiple times.

As a bonus if we start using pytest fixtures this will fix the
problems we were gonna hit running the flake finder on those.

I also moved these to @tryfirst so that randomization happens after
the whole list is made.

Test Plan: Ran --flake-finder on test/test_trace.py it would fail before, now it always passes.