Fix flakefinder not duplicating tests in derived classes that used runTest.
Summary: When using unittest.TestCase.runTest and making subclasses of
it the flake finder would end up adding the "duplicated" mark to the
function object.  If the function object is on the base class then it
would make the super-classes test not get duplicated.

The solution is not to mark the items.  We have a list, we can depend
that items in the list won't be duplicated already.

Test Plan: Added a test case.