Handle latest pytest output
The latest versions of `pytest` (3.3.0+) have some changes to the output
that break the flakefinder tests:

- `collected X items` is now properly pluralized
- By default, pytest output now adds the test progress percent to the
  end of each line.

Also a tweak to a comment in `tox.ini` that was breaking for `tox` on
Python 3.6.

Also, 3.3.0+ will now print out warning logs indicating
`metafunc.addcall` is deprecated and will be removed in 4.0+. This
breaks one of the tests that check the summary output. So patch that for
now. I have another fix to handle the deprecation coming up. Just needed
working tests first!